Make some utility functions for tests
